requirements.txt
google-api-python-client==2.116.0
google-cloud-secret-manager==2.18.1
requests==2.31.0
fastavro==1.9.3
confluent-kafka==2.3.0